pyproject.toml
[project]
name = "roadmaptool"
version = "0.1.0"
description = "Browser-based Roadmap/Gantt chart tool"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "ruamel.yaml>=0.18.6",
    "pydantic>=2.10.0",
]

[project.scripts]
app = "roadmaptool.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/roadmaptool"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]
